old/requirements.txt
pyserial==3.5
RPi.GPIO==0.7.1
ffpyplayer==4.5.2